Fixed naming and access modifiers
